feat: implement digital asset management with CRUD operations and private file handling

+++ database/migrations/2025_05_22_102608_create_telescope_entries_table.php
@@ -6,8 +6,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Get the migration connection name.
      */
@@ -23,39 +22,43 @@ return new class extends Migration
     {
         $schema = Schema::connection($this->getConnection());
 
-        $schema->create('telescope_entries', function (Blueprint $table) {
-            $table->bigIncrements('sequence');
-            $table->uuid('uuid');
-            $table->uuid('batch_id');
-            $table->string('family_hash')->nullable();
-            $table->boolean('should_display_on_index')->default(true);
-            $table->string('type', 20);
-            $table->longText('content');
-            $table->dateTime('created_at')->nullable();
+        if (!$schema->hasTable('telescope_entries')) {
+            $schema->create('telescope_entries', function (Blueprint $table) {
+                $table->bigIncrements('sequence');
+                $table->uuid('uuid');
+                $table->uuid('batch_id');
+                $table->string('family_hash')->nullable();
+                $table->boolean('should_display_on_index')->default(true);
+                $table->string('type', 20);
+                $table->longText('content');
+                $table->dateTime('created_at')->nullable();
 
-            $table->unique('uuid');
-            $table->index('batch_id');
-            $table->index('family_hash');
-            $table->index('created_at');
-            $table->index(['type', 'should_display_on_index']);
-        });
+                $table->unique('uuid');
+                $table->index('batch_id');
+                $table->index('family_hash');
+                $table->index('created_at');
+                $table->index(['type', 'should_display_on_index']);
+            });
+        }
+        if (!$schema->hasTable('telescope_entries_tags')) {
+            $schema->create('telescope_entries_tags', function (Blueprint $table) {
+                $table->uuid('entry_uuid');
+                $table->string('tag');
 
-        $schema->create('telescope_entries_tags', function (Blueprint $table) {
-            $table->uuid('entry_uuid');
-            $table->string('tag');
+                $table->primary(['entry_uuid', 'tag']);
+                $table->index('tag');
 
-            $table->primary(['entry_uuid', 'tag']);
-            $table->index('tag');
-
-            $table->foreign('entry_uuid')
-                ->references('uuid')
-                ->on('telescope_entries')
-                ->onDelete('cascade');
-        });
-
-        $schema->create('telescope_monitoring', function (Blueprint $table) {
-            $table->string('tag')->primary();
-        });
+                $table->foreign('entry_uuid')
+                    ->references('uuid')
+                    ->on('telescope_entries')
+                    ->onDelete('cascade');
+            });
+        }
+        if (!$schema->hasTable('telescope_monitoring')) {
+            $schema->create('telescope_monitoring', function (Blueprint $table) {
+                $table->string('tag')->primary();
+            });
+        }
     }
 
     /**
